+++ test/core/edge.sv
@@ -1,0 +1,8 @@
+module mod(input clk);
+    always @(posedge clk)
+        $display($time, "posedge");
+    always @(negedge clk)
+        $display($time, "negedge");
+    always @(edge clk)
+        $display($time, "edge");
+endmodule