+++ test/core/edge_tb.v
@@ -1,0 +1,9 @@
+module top;
+    reg clk;
+    initial begin
+        clk = 0;
+        repeat (10)
+            #5 clk = ~clk;
+    end
+    mod m(clk);
+endmodule