+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added support for excluding the conversion of unbased unsized literals (e.g.,
   `'1`, `'x`) via `--exclude UnbasedUniszed`
 * Added support for enumerated type ranges (e.g., `enum { X[3:5] }`)
+* Added support for the SystemVerilog `edge` event
 * Added support for passing through DPI imports and exports
 * Added support for passing through functions with output ports